pyproject.toml
[tool.poetry]
name = "vk-users"
version = "0.1.0"
description = "App that collects vk users data"
authors = ["Elizaveta"]
readme = "README.md"
packages = [{include = "vk_users"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
flask = "^3.0.2"
gunicorn = "^21.2.0"
datetime = "^5.4"
tqdm = "^4.66.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
